% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafilter-package.R
\docType{package}
\name{datafilter-package}
\alias{datafilter}
\alias{datafilter-package}
\title{datafilter: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Darren Kidney \email{darrenkidney@gmail.com}

}
\keyword{internal}
