% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFilter.R
\name{DataFilter}
\alias{DataFilter}
\title{R6 Class for data frame filtering}
\value{
todo
}
\description{
todo
}
\examples{

data = load_example_data()
df = DataFilter$new(data)
df$profile
df$drop(data) |> head()

data = mtcars
df = DataFilter$new(data)
df$profile
df$drop(data) |> head()

data = airquality
df = DataFilter$new(data)
df$profile
df$drop(data) |> head()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{profile}}{column profile}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataFilter-new}{\code{DataFilter$new()}}
\item \href{#method-DataFilter-drop}{\code{DataFilter$drop()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataFilter-new}{}}}
\subsection{Method \code{new()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFilter$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFilter-drop"></a>}}
\if{latex}{\out{\hypertarget{method-DataFilter-drop}{}}}
\subsection{Method \code{drop()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFilter$drop(
  data,
  min_n_unique = 2,
  max_p_na = 0.9,
  min_var_numeric = 1e-08,
  min_var_binary = 0.01,
  max_p_mode = 0.99,
  drop_empty_rows = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{todo}

\item{\code{min_n_unique}}{todo}

\item{\code{max_p_na}}{todo}

\item{\code{min_var_numeric}}{todo}

\item{\code{min_var_binary}}{todo}

\item{\code{max_p_mode}}{todo}

\item{\code{drop_empty_rows}}{todo}
}
\if{html}{\out{</div>}}
}
}
}
