% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_columns.R
\name{profile_columns}
\alias{profile_columns}
\title{Column profiler}
\usage{
profile_columns(data, p_mode_min = 0.5, p_mode_max_card = 30)
}
\arguments{
\item{data}{a data.frame or tibble}

\item{p_mode_min}{todo}

\item{p_mode_max_card}{todo}
}
\value{
A tibble with rows representing the columns of \code{data} and columns representing
}
\description{
A short description...
}
\examples{
(df = load_example_data())
(df_profile = df |> profile_columns())
df |> drop_columns(df_profile, dryrun=TRUE)
df |> drop_columns(df_profile)

(df = mtcars |> dplyr::as_tibble())
(df_profile = df |> profile_columns())
df |> drop_columns(df_profile, dryrun=TRUE)
df |> drop_columns(df_profile)
}
