% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{drop_columns}
\alias{drop_columns}
\title{Title}
\usage{
drop_columns(
  data,
  profile = NULL,
  dryrun = TRUE,
  min_n_unique = 2,
  max_p_na = 0.9,
  min_var_numeric = 1e-08,
  min_var_binary = 0.01,
  max_p_mode = 0.99,
  ...
)
}
\arguments{
\item{data}{todo}

\item{profile}{todo}

\item{dryrun}{todo}

\item{min_n_unique}{todo}

\item{max_p_na}{todo}

\item{min_var_numeric}{todo}

\item{min_var_binary}{todo}

\item{max_p_mode}{todo}

\item{...}{todo}
}
\description{
Title
}
\examples{
# todo
}
