% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_columns}
\alias{profile_columns}
\title{Column profiler}
\usage{
profile_columns(
  data,
  p_mode_min = 0.5,
  p_mode_max_card = 30,
  flag_high_card_cat = 30,
  flag_low_card_num = 10
)
}
\arguments{
\item{data}{a data.frame or tibble}

\item{p_mode_min}{todo}

\item{p_mode_max_card}{todo}

\item{flag_high_card_cat}{todo}

\item{flag_low_card_num}{todo}
}
\value{
A tibble with rows representing the columns of \code{data} and columns representing
}
\description{
A short description...
}
\examples{
(example_data = load_example_data())
(example_data_profile = example_data |> profile_columns())
example_data |> drop_columns(example_data_profile, dryrun=TRUE)

(mtcars_profile = mtcars |> profile_columns())
mtcars |> drop_columns(mtcars_profile, dryrun=TRUE)
}
